fix(kadena-cli): add logger for command and config
